+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Remove support for PHP7
 - Remove support for PHPUnit6 and PHPUnit7
+- Remove deprecated `Phake::partMock()`